requirements.txt
absl-py==0.7.1
antiorm==1.2.1
asgiref==3.2.3
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
cachetools==4.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
db==0.1.1
db-sqlite3==0.0.1
decorator==4.3.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.5
django-ckeditor==5.8.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mysql==3.3.0
django-tinymce==2.8.0
django-tinymce4-lite==1.8.0
django-utils==0.0.2
dnspython==1.16.0
entrypoints==0.3
gast==0.2.2
gcloud==0.18.3
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-pasta==0.1.7
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.22.0
gunicorn==20.0.4
h5py==2.9.0
httplib2==0.17.3
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
joblib==0.13.2
jsmin==2.2.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterthemes==0.20.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lesscpy==0.13.0
manage.py==0.2.10
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mnist==0.2.2
mrcnn==0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
nltk==3.4.4
notebook==5.7.6
numpy==1.16.4
oauth2client==4.1.3
opencv-python==4.1.0.25
pandas==0.24.2
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.3.4
pickleshare==0.7.5
Pillow==6.0.0
pipenv==2018.11.26
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.3.1
pyparsing==2.4.0
PyQRCode==1.2.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.22.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
tornado==6.0.1
traitlets==4.3.2
urllib3==1.25.3
virtualenv==16.4.3
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
whitenoise==5.0.1
widgetsnbextension==3.4.2
wrapt==1.11.2
wxPython==4.0.6
xlrd==1.2.0
